実装場所をenemyServiceへ移動
