+++ backend/src/rooms/GameRoom.ts
@@ -4,19 +4,6 @@ import Matter from 'matter-js';
 import * as Constants from '../constants/constants';
 import dropWalls from '../game_engine/services/dropWallService';
 import PlacementObjectInterface from '../interfaces/placement_object';
-import {
-  directMovableMap,
-  getDirectMovableMapIfBombSet,
-  getHighestPriorityTile,
-  influenceToOtherTile,
-  isSelfDie,
-  normalizeDimension,
-  numberOfDestroyableBlock,
-  reverseNormalizeDimension,
-  searchPath,
-  sumOfProductsSynthesis,
-  treatLevelMapByBomb,
-} from '../utils/calc';
 import GameQueue from '../utils/gameQueue';
 import GameEngine from './GameEngine';
 import Block from './schema/Block';
@@ -24,7 +11,6 @@ import { Bomb } from './schema/Bomb';
 import Enemy from './schema/Enemy';
 import GameRoomState from './schema/GameRoomState';
 import Item from './schema/Item';
-import Player from './schema/Player';
 
 export default class GameRoom extends Room<GameRoomState> {
   engine!: GameEngine;
@@ -262,229 +248,10 @@ export default class GameRoom extends Room<GameRoomState> {
     }
   }
 
-  // 爆弾を設置する
-  enemySetBomb(directMoveMap: number[][], highPriorityForBlastRadiusMap: number[][], enemy: Enemy) {
-    // もし爆弾を置いたらどうなるか？のマップを作成する
-    const mapIfSetBomb = getDirectMovableMapIfBombSet(
-      directMoveMap,
-      highPriorityForBlastRadiusMap,
-      enemy.x,
-      enemy.y,
-      enemy.bombStrength
-    );
-
-    // console.log('mapIfSetBomb', mapIfSetBomb);
-    if (!enemy.canSetBomb()) return;
-    if (isSelfDie(directMoveMap, mapIfSetBomb, enemy, true)) return;
-    // console.log('mapIfSetBomb', mapIfSetBomb);
-    // enemy.setGoal(getClosestAvailablePoint(mapIfSetBomb, {x, y: enemy.getTilePosition()});
-    this.engine.bombService.enqueueBomb(enemy as Player);
-  }
-
-  enemyHandler() {
+  // 敵の移動を行う
+  private enemyHandler() {
     if (!this.dummyFlag) return;
     if (!this.state.gameState.isPlaying()) return;
-
-    // const DimensionalMap = this.engine.getDimensionalMap((bodies) => bodies);
-
-    // 爆弾の影響度マップを作成する
-    const futureBlastMap = treatLevelMapByBomb(
-      this.engine.getDimensionalMap(this.engine.getHighestPriorityFromBodies),
-      this.state.hasBomb(this.engine.getDimensionalMap((bodies) => bodies))
-    );
-
-    // 爆風の位置から死亡マップを作成する
-    const deathMap = reverseNormalizeDimension(
-      normalizeDimension(influenceToOtherTile(futureBlastMap))
-    );
-
-    // 移動できるマス(ボムでブロックを破壊できる場所も含む)のマップを作成する
-    // この時将来発生する爆風のマスは0にする
-    const checkMovableDimensionalMap = this.engine.getDimensionalMap(this.engine.checkMovable);
-    // .map((row, r) => row.map((col, c) => (futureBlastMap[r][c] === 1 ? 0 : col)));
-
-    const movableMap = normalizeDimension(checkMovableDimensionalMap);
-    // 爆風の範囲を測定するために使うマップを作成する
-    const highPriorityForBlastRadiusMap = this.engine.getDimensionalMap(
-      this.engine.getHighestPriorityFromBodies
-    );
-    // アイテムのマップ
-    const itemMap = normalizeDimension(
-      influenceToOtherTile(this.engine.getDimensionalMap(this.engine.HasItem))
-    );
-    // ブロックのマップ
-    const blockMap = normalizeDimension(this.engine.getDimensionalMap(this.engine.HasBlock));
-
-    // 爆風のマップ
-    const blastMap = this.engine.getDimensionalMap(this.engine.HasBlast);
-    // if (deathMap.flat().filter((v: number) => v < 1).length > 0) console.log('deathMap', deathMap);
-
-    // 特定のマスの周囲のマスにどの程度空きマスがないか
-    // const notFreeSpaceMap = reverseNormalizeDimension(
-    //   normalizeDimension(treatLevelByFreeSpace(checkMovableDimensionalMap, 3))
-    // );
-
-    for (let i = 0; i < Constants.DEBUG_DEFAULT_ENEMY_COUNT; i++) {
-      const player = this.state.getPlayer(`enemy-${i}`);
-      if (player === undefined) continue;
-
-      const enemy = player as Enemy;
-      if (enemy.isDead()) continue;
-
-      const { x: enemyX, y: enemyY } = enemy.getTilePosition();
-
-      // 直接移動できるマスのマップを作成する
-      const MoveCountMap = directMovableMap(checkMovableDimensionalMap, enemyX, enemyY);
-      const directMoveMap = MoveCountMap.map((row, r) =>
-        row.map((v, c) => (row[c] = v === Infinity ? 0 : 1))
-      );
-
-      // 一度に破壊できるブロックが多い場所のマップを作成する
-      const goodBombPlaceMap = influenceToOtherTile(
-        normalizeDimension(
-          numberOfDestroyableBlock(directMoveMap, blockMap, highPriorityForBlastRadiusMap, enemy)
-        )
-      );
-
-      // 影響度マップを作成する
-      const impactMap = sumOfProductsSynthesis(movableMap, [
-        // 爆弾、爆風の影響度マップ
-        {
-          dimensionalMap: deathMap,
-          ratio: Constants.ENEMY_EVALUATION_RATIO_BOMB,
-        },
-        // 現在地点からの距離の影響度マップ
-        {
-          dimensionalMap: MoveCountMap,
-          ratio: Constants.ENEMY_EVALUATION_RATIO_NEAREST,
-        },
-        // {
-        //   // dimensionalMap: movableMap,
-        //   dimensionalMap: notFreeSpaceMap,
-        //   ratio: Constants.ENEMY_EVALUATION_RATIO_FREE_SPACE,
-        // },
-        // アイテムの影響度マップ
-        {
-          dimensionalMap: itemMap,
-          ratio: Constants.ENEMY_EVALUATION_RATIO_ITEM,
-        },
-        // 破壊できるブロックの数の影響度マップ
-        {
-          dimensionalMap: goodBombPlaceMap,
-          ratio: Constants.ENEMY_EVALUATION_RATIO_GOOD_BOMB_PLACE,
-        },
-      ]);
-
-      // 影響度マップに対して、今移動できるマスのみを残す
-      const impactMapIsMovable = impactMap.map((row, i) =>
-        row.map((v, j) => v * directMoveMap[i][j])
-      );
-      // if (deathMap.flat().filter((v: number) => v < 1).length > 0) {
-      // console.log('directMoveMap', directMoveMap);
-      //   console.log('impactMap', impactMap);
-      // console.log('impactMapIsMovable', impactMapIsMovable);
-      // }
-
-      // console.log('directMoveMap', directMoveMap);
-      // console.log('impactMap', impactMap);
-      // if (deathMap.flat().filter((v: number) => v < 1).length > 0) {
-      //   console.log('impactMap', impactMap);
-      //   console.log('impactMapIsMovable', impactMapIsMovable);
-      //   console.log('directMoveMap', directMoveMap);
-      // }
-
-      // impactMapIsMovable から、現在地点の周囲のマスの中で、最も影響度が高いマスを取得する
-      const { x, y } = getHighestPriorityTile(impactMapIsMovable, enemyX, enemyY);
-      // const { x, y } = getHighestPriorityTileSurround(impactMapIsMovable, enemyX, enemyY);
-
-      // 最終的に移動するマスを決定する
-      enemy.setGoal(x, y);
-
-      // ゴールに着くまで移動
-
-      if (!enemy.isMovedToGoal()) {
-        // 最短の移動経路を取得する
-        const moveList = searchPath(
-          enemy.getTilePosition(),
-          enemy.getGoalTilePosition(),
-          directMoveMap
-        );
-
-        // console.log(
-        //   'now:',
-        //   enemy.getTilePosition(),
-        //   'next:',
-        //   enemy.getNextTilePosition(),
-        //   'goal:',
-        //   enemy.getGoalTilePosition()
-        // );
-        // console.log('impactMap', impactMap);
-        // console.log('impactMapIsMovable', impactMapIsMovable);
-
-        // 移動経路の結果が 0 の場合は、すでに目的にいるので何もしない
-        if (moveList.length === 0) continue;
-        // moveList.forEach((v) => console.log('moveList', v));
-
-        // 次に移動するマスの安全度が1/2以下 or 爆風があるマスなるなら動かない
-        if (
-          impactMapIsMovable[moveList[1][1]][moveList[1][0]] <=
-            impactMapIsMovable[enemyY][enemyX] * 0.5 ||
-          blastMap[moveList[1][1]][moveList[1][0]] === 1
-        ) {
-          // console.log('stop', enemy.x, enemy.y);
-          //　キューに移動が残ってると止めても移動するので、キューをクリアする
-          this.engine.enemyService.stop(enemy);
-          continue;
-        }
-
-        enemy.setNext(moveList[1][0], moveList[1][1]);
-
-        // 次に移動するマスの脅威度が 0.3 以下の場合は、移動しない
-        // if (impactMapIsMovable[moveList[1][1]][moveList[1][0]] < 0.1) continue;
-
-        // if (deathMap.flat().filter((v: number) => v < 1).length > 0) console.log('x, y', x, y);
-        const key = enemy.moveToDirection();
-        const data = {
-          player: enemy,
-          inputPayload: {
-            up: key.up,
-            down: key.down,
-            left: key.left,
-            right: key.right,
-          },
-          isInput: true,
-        };
-        enemy.inputQueue.push(data);
-        // console.log('set bomb');
-      } else {
-        // 爆風の位置を含めて、マス目の安全度が 0.5 以上のマスを移動可能とみなす
-        const safeTileRate = 0.5;
-        const directMoveMapIncludeBlast = impactMapIsMovable.map((row, i) =>
-          row.map((v, j) => (v >= safeTileRate ? 1 : 0))
-        );
-        this.enemySetBomb(directMoveMapIncludeBlast, highPriorityForBlastRadiusMap, enemy);
-        // console.log(
-        //   'now:',
-        //   enemy.getTilePosition()
-        //   // 'next:',
-        //   // enemy.getNextTilePosition(),
-        //   // 'goal:',
-        //   // enemy.getGoalTilePosition()
-        // );
-      }
-      // }
-
-      // const surroundTiles = enemy.getSurroundingTiles(movableMap);
-
-      // for (const [key, value] of Object.entries(surroundTiles).sort(() => Math.random() - 0.5)) {
-      //   if (this.engine.isMovable(movableMap[value.y][value.x])) {
-      //     enemy.moveToNextTile(value.x, value.y);
-      //   }
-      // }
-
-      // 爆弾設置
-      // if (!player.canSetBomb()) return;
-      // this.state.getBombToCreateQueue().enqueue(this.state.createBomb(player));
-    }
+    this.engine.enemyService.calcAdjustablePosition();
   }
 }
